# backend/TeamFlow.API/DTOs/Auth/LoginDto.cs
using System.ComponentModel.DataAnnotations;

namespace TeamFlow.API.DTOs.Auth
{
    public class LoginDto
    {
        [Required(ErrorMessage = "Email jest wymagany")]
        [EmailAddress(ErrorMessage = "Nieprawidłowy format email")]
        public string Email { get; set; } = string.Empty;

        [Required(ErrorMessage = "Hasło jest wymagane")]
        public string Password { get; set; } = string.Empty;
    }
}